Transformer/requirements.txt
# Core ML/NLP
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# MLOps
mlflow>=2.5.0
dvc>=3.0.0

# API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Monitoring
prometheus-client>=0.17.0
prometheus-fastapi-instrumentator>=6.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.24.0

# Code Quality
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.4.0

# Security
bandit>=1.7.5

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0